Fix SHA256 download: follow redirects (#132)
fixes #131

It looks like something in the hosting of dl.gitea.io has changed (@Owners please confirm?), so that requests are redirected to `https://storage.gitea.io/gitea-artifacts/...`
The binary downloads are not affected, as `wget` is used which follows redirects by default.

edit: Is the redirect URI the new canonical location? In that case it would make more sense to update the base URL instead..

Reviewed-on: https://gitea.com/gitea/homebrew-gitea/pulls/132
Reviewed-by: techknowlogick <techknowlogick@gitea.io>
Co-authored-by: Norwin <noerw@noreply.gitea.io>
Co-committed-by: Norwin <noerw@noreply.gitea.io>